(PC-27701) refactor(venumap): add a directory for the feature venue map
