+++ src/features/venuemap/components/VenueMapBlock/VenueMapBackground.ts
